Enable a number inductor of tests on CPU
There were many test that their `_cuda` variants were not running on
cuda. I fixed a few of these, but I'm sure there's plenty more.
It'd be great to have a way to test that we're indeed compiling
something in these tests, but I don't know how to do this off the top of
my head.

ghstack-source-id: af4dbd43fe63de6d5f81fd3fc33dd09daa0d8de7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/107465